Move Swagger config in AppConfig, hide ErrorBasicController from Swagger
