+++ src/main/java/com/school/project/configuration/AppConfiguration.java
@@ -1,8 +1,17 @@
 package com.school.project.configuration;
 
+import com.google.common.base.Predicates;
+import org.hibernate.event.spi.PostLoadEvent;
 import org.modelmapper.ModelMapper;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import springfox.documentation.builders.PathSelectors;
+import springfox.documentation.builders.RequestHandlerSelectors;
+import springfox.documentation.service.ApiInfo;
+import springfox.documentation.spi.DocumentationType;
+import springfox.documentation.spring.web.plugins.Docket;
+
+import javax.annotation.PostConstruct;
 
 @Configuration
 public class AppConfiguration {
@@ -11,5 +20,17 @@ public class AppConfiguration {
     public ModelMapper modelMapper() {
         ModelMapper modelMapper = new ModelMapper();
         return modelMapper;
+    }
+
+    @Bean
+    public Docket api() {
+        return new Docket(DocumentationType.SWAGGER_2)
+                .select()
+                .apis(RequestHandlerSelectors.any())
+                .paths(Predicates.not(PathSelectors.regex("/error.*")))
+                .build();
+    }
+    private ApiInfo apiEndPointsInfo() {
+        return ApiInfo.DEFAULT;
     }
 }
